Favorite 관련 기능 분리 및 fetch join 수정
