+++ src/main/java/com/example/ca4u_backend/domain/club/ClubService.java
@@ -20,10 +20,6 @@ import java.util.List;
 @Service
 public class ClubService {
     private final ClubRepository clubRepository;
-    private final FavoriteRepository favoriteRepository;
-    private final FavoriteHistoryRepository FavoriteHistoryRepository;
-    private final UserRepository userRepository;
-
 
     public ClubReponseDto getClubSpec(long clubId) {
         Club club = clubRepository.findById(clubId).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 클럽입니다."));
@@ -38,33 +34,5 @@ public class ClubService {
     public List<ClubReponseDto> getClubsBySearch(String search) {
         List<Club> clubList = clubRepository.findByClubNmContainingOrBriefDescriptionContaining(search, search);
         return clubList.stream().map(ClubReponseDto::of).toList();
-    }
-
-    @Transactional
-    public String toggleClubFavoriteStatus(long clubId, long userId) {
-        Club club = clubRepository.findById(clubId).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 클럽입니다."));
-        User user = userRepository.findById(userId).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 사용자입니다."));
-
-        String userName = user.getUsername();
-        String clubName = club.getClubNm();
-
-        if (favoriteRepository.existsByUserAndClub(user, club)) {
-            favoriteRepository.deleteByUserAndClub(user, club);
-            //이력 추가
-            FavoriteHistoryRepository.save(new FavoriteHistory(user, club, ActionType.REMOVE));
-            return userName + "님이 " + clubName + "를 즐겨찾기 목록에서 삭제했습니다.";
-
-        } else {
-            favoriteRepository.save(new Favorite(user, club));
-            //이력 추가
-            FavoriteHistoryRepository.save(new FavoriteHistory(user, club, ActionType.ADD));
-            return userName + "님이 " + clubName + "를 즐겨찾기 목록에 추가했습니다.";
-        }
-    }
-
-    public List<ClubReponseDto> getFavoriteClubsByUserId(Long userId) {
-        User user = userRepository.findById(userId).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 사용자입니다."));
-        List<Favorite> favoriteList = favoriteRepository.findByUser(user);
-        return favoriteList.stream().map(Favorite::getClub).map(ClubReponseDto::of).toList();
     }
 }
